updates: React to network changes
The Automatic Updates panel should represent the current
state of the network. That means that, whenever an external
tool changes the automatic updates settings, the panel must
react to those changes and always display the most updated
value.

This is currently not the case, though, since it is not
reacting to the current connection changes, potentially
showing obsolete values.

Fix that by properly storing and connecting to the connection's
changed signal, and updating the settings when the signal is
fired.

https://phabricator.endlessm.com/T21823