updates: Introduce Automatic Updates panel
The Automatic Updates panel is a new panel that manages two
different settings:

 1. Whether updates should happen automatically or not
 2. A period where updates are allowed to happen

The first setting is implemented on a per-connection fashion,
using NetworkManager's ability to store user data on specific
connections. In this case, the property being set is:

  connection.allow-downloads-when-metered

The second setting is slightly trickier to implement, and uses
libmogwai-tariff to load and store a simple tariff. This tariff
has the following layout:

  i. Base tariff, downloads are forbidden, spans from 0 to forever.
 ii. Recurrent tariff, downloads are completely allowed, spans
     from [start hour] to [end hour]. If [start hour] >= [end hour],
     then [end hour] is contained in the following day.

This tariff schedule, when changed, is saved every 2 seconds, to
avoid abusing I/O.

https://phabricator.endlessm.com/T20818

Rebase-3.32 note: Switch build system to meson for updates panel
(T25338)