Re-apply intentional downstream translations
These seem to be intentional changes vs. upstream, which didn't get
pulled in via the normal process because the source strings had changed.

Many of our intentional translations seem to have been moved upstream so
this patch is much smaller than in the previous iteration.

I verified that the intentional translations re-introduced in
17f7914 during a previous rebase are also reintroduced here.

https://phabricator.endlessm.com/T26907